requirements.txt
absl-py==0.7.1
aiofiles==0.5.0
aiohttp==3.7.4.post0
aioice==0.7.5
aiortc==1.2.0
alembic==1.6.5
altgraph==0.17
amqp==5.0.2
aniso8601==7.0.0
appdirs==1.4.4
APScheduler==3.6.3
asgiref==3.3.0
astor==0.8.0
astroid==2.4.1
async-exit-stack==1.0.1
async-generator==1.10
async-timeout==3.0.1
attrs==20.3.0
autobahn==21.3.1
Automat==20.2.0
backcall==0.1.0
bcrypt==3.2.0
CacheControl==0.12.6
cachetools==4.2.2
constantly==15.1.0
crc32c==2.2
cryptography==3.4.7
decorator==4.4.0
defusedxml==0.6.0
Django==3.1.3
django-cors-headers==3.5.0
djangorestframework==3.12.1
djangorestframework-simplejwt==4.6.0
entrypoints==0.3
HTMLParser==0.0.2
jedi==0.17.0
Jinja2==2.11.2
jsonschema==3.0.1
multidict==5.1.0
mysql==0.0.2
mysql-connector==2.2.9
mysql-connector-python==8.0.21
mysqlclient==2.0.3
passlib==1.7.4
pydantic==1.8.1
PyMySQL==0.10.0
PyYAML==5.4.1
requests==2.25.1
uritemplate==3.0.1
urllib3==1.26.6
Werkzeug==0.15.4
yarl==1.6.3
